post migration cleanup
