+++ package.json
@@ -17,9 +17,7 @@
     "db:prod:download": "src/scripts/downloadDatabase.sh",
     "generateCanopy": "ts-node --project tsconfig.script.json src/scripts/generateCanopy.ts",
     "tursoMigration": "src/scripts/tursoMigration.sh",
-    "downloadCloudAssets": "ts-node --project tsconfig.script.json src/scripts/downloadCloudAssets.ts",
-    "backupLocalExperienceAssets": "ts-node --project tsconfig.script.json src/scripts/backupLocalExperienceAssets.ts",
-    "backup": "ts-node --project tsconfig.script.json src/scripts/backupLocalExperienceAssets.ts",
+    "downloadCloudAudio": "ts-node --project tsconfig.script.json src/scripts/downloadCloudAssets.ts",
     "unityTestServer": "ts-node --project tsconfig.script.json src/scripts/unityTestServer.ts",
     "controllerServer": "ts-node --project tsconfig.script.json src/scripts/controllerServer.ts",
     "generatePattern": "ts-node --project tsconfig.script.json src/scripts/generatePattern.ts"
