--- src/scripts/backupLocalExperienceAssets.ts
@@ -1,8 +1,0 @@
-import { backupExperiences } from "../scripts/assetManagement";
-
-const main = async () => {
-  console.log("backing up experience assets...");
-  await backupExperiences();
-};
-
-main();
